requirements.txt
python-pptx>=0.6.21
selenium>=4.15.0
webdriver-manager>=4.0.0
beautifulsoup4>=4.12.0
matplotlib>=3.8.0
pillow>=10.3.0
tqdm>=4.66.0
colorama>=0.4.6
lxml>=5.0.0
python-docx>=0.8.11
requests>=2.31.0
comtypes>=1.4.0

